+++ src/test/java/io/kuzzle/test/security/KuzzleSecurity/getProfileTest.java
@@ -44,7 +44,7 @@ public class getProfileTest {
   }
 
   @Test
-  public void testgetProfileValidResponse() throws JSONException {
+  public void testgetProfileValidResponseWithoutHydrate() throws JSONException {
     doAnswer(new Answer() {
       @Override
       public Object answer(InvocationOnMock invocation) throws Throwable {
@@ -52,9 +52,69 @@ public class getProfileTest {
           "{" +
             "\"result\": {" +
             "\"_id\": \"foobar\"," +
-            "\"_source\": {}" +
+            "\"_source\": {" +
+              "\"roles\": [{\"_id\": \"foo\", \"_source\": {\"restrictedTo\": []," +
+              "\"allowInternalIndex\": true}" +
+              "}" +
+              "]" +
             "}" +
-            "}");
+            "}" +
+              "}");
+
+        ((OnQueryDoneListener) invocation.getArguments()[3]).onSuccess(response);
+        ((OnQueryDoneListener) invocation.getArguments()[3]).onError(new JSONObject().put("error", "stub"));
+        return null;
+      }
+    }).when(kuzzle).query(any(io.kuzzle.sdk.core.Kuzzle.QueryArgs.class), any(JSONObject.class), any(KuzzleOptions.class), any(OnQueryDoneListener.class));
+
+    KuzzleOptions options = new KuzzleOptions();
+    options.setHydrate(false);
+
+    kuzzleSecurity.getProfile("foobar", options, new KuzzleResponseListener<KuzzleProfile>() {
+      @Override
+      public void onSuccess(KuzzleProfile response) {
+        assertEquals(response.id, "foobar");
+        assertEquals(true, response.getRoles()[0].getContent().has("restrictedTo"));
+        try {
+          assertEquals(true, response.getRoles()[0].getContent().getBoolean("allowInternalIndex"));
+        } catch (JSONException e) {
+          e.printStackTrace();
+        }
+      }
+
+      @Override
+      public void onError(JSONObject error) {
+        try {
+          assertEquals(error.getString("error"), "stub");
+        } catch (JSONException e) {
+          throw new RuntimeException(e);
+        }
+      }
+    });
+
+    ArgumentCaptor argument = ArgumentCaptor.forClass(io.kuzzle.sdk.core.Kuzzle.QueryArgs.class);
+    verify(kuzzle, times(1)).query((io.kuzzle.sdk.core.Kuzzle.QueryArgs) argument.capture(), any(JSONObject.class), any(KuzzleOptions.class), any(OnQueryDoneListener.class));
+    assertEquals(((io.kuzzle.sdk.core.Kuzzle.QueryArgs) argument.getValue()).controller, "security");
+    assertEquals(((io.kuzzle.sdk.core.Kuzzle.QueryArgs) argument.getValue()).action, "getProfile");
+  }
+
+  @Test
+  public void testgetProfileValidResponseWithHydrate() throws JSONException {
+    doAnswer(new Answer() {
+      @Override
+      public Object answer(InvocationOnMock invocation) throws Throwable {
+        JSONObject response = new JSONObject(
+            "{" +
+                "\"result\": {" +
+                "\"_id\": \"foobar\"," +
+                "\"_source\": {" +
+                "\"roles\": [{\"_id\": \"foo\", \"_source\": {\"foo\": []," +
+                "}" +
+                "}" +
+                "]" +
+                "}" +
+                "}" +
+                "}");
 
         ((OnQueryDoneListener) invocation.getArguments()[3]).onSuccess(response);
         ((OnQueryDoneListener) invocation.getArguments()[3]).onError(new JSONObject().put("error", "stub"));
@@ -66,6 +126,7 @@ public class getProfileTest {
       @Override
       public void onSuccess(KuzzleProfile response) {
         assertEquals(response.id, "foobar");
+        assertEquals(true, response.getRoles()[0].getContent().has("foo"));
       }
 
       @Override